backend/requirements.txt
python-dotenv>=1.2,<2.0
strands-agents>=1.35,<2.0
strands-agents-tools>=0.4,<1.0
ollama>=0.6,<1.0
pydantic>=2.12,<3.0
httpx>=0.28,<1.0
PyYAML>=6.0,<7.0
jinja2>=3.1,<4.0
beautifulsoup4>=4.14,<5.0
fastapi>=0.135,<1.0
python-multipart>=0.0.26,<1.0
uvicorn[standard]>=0.44,<1.0
slack-sdk>=3.41,<4
cryptography>=46.0,<47.0
pytest>=9.0,<10.0
pytest-cov>=7.1,<8.0
temporalio>=1.25,<2.0
APScheduler>=3.11,<4.0
pytz>=2026.1
playwright>=1.58,<2.0
psycopg[binary]>=3.3,<4.0
yfinance>=1.2,<2.0
pandas>=3.0,<4.0
numpy>=2.4,<3.0
prometheus-fastapi-instrumentator>=7.1,<8.0
opentelemetry-api>=1.41,<2.0
opentelemetry-sdk>=1.41,<2.0
opentelemetry-exporter-otlp-proto-http>=1.41,<2.0
opentelemetry-exporter-otlp-proto-grpc>=1.41,<2.0
opentelemetry-instrumentation-fastapi>=0.62b0,<1.0
opentelemetry-instrumentation-httpx>=0.62b0,<1.0
opentelemetry-instrumentation-logging>=0.62b0,<1.0
